% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{PRISMA2Dos1cor}
\alias{PRISMA2Dos1cor}
\title{PRISMA2Dos1cor}
\usage{
PRISMA2Dos1cor(rastList, verbose = F)
}
\arguments{
\item{rastList}{list: output from PRISMA2rast function}

\item{verbose}{logical: lot's of messages if True. Defaults to False.}
}
\value{
A list of terra::rast objects with values from 0 to 1 representing
corrected digital numbers of surface reflectance.
\itemize{
\item swir data cube
\item vnir data cube
}
}
\description{
Atmospheric correction of PRISMA data using DOS1 method (see Chavez 1996)
}
\examples{
### filepath<-"/archivio/shared/geodati/raster/OPTICAL/PRISMA/"
### filename<-"PRS_L2D_STD_20200418101701_20200418101706_0001.he5"
### fn <- PRISMA2rast( file.path(filepath, filename), verbose=T)
}
