% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{PRISMA2ext}
\alias{PRISMA2ext}
\title{PRISMA2ext}
\usage{
PRISMA2ext(input, verbose = F)
}
\arguments{
\item{input}{character: text of the filepath to HDF PRISMA dataset}

\item{verbose}{logical: lot's of messages if True. Defaults to False.}
}
\value{
A numeric vector (length=4; order= xmin, xmax, ymin, ymax)
}
\description{
Extracts extension of PRISMA image. It can be used
}
\examples{
### filepath<-"/archivio/shared/geodati/raster/OPTICAL/PRISMA/"
### filename<-"PRS_L2D_STD_20200418101701_20200418101706_0001.he5"
### fn <- PRISMA2ext( file.path(filepath, filename), verbose=T)
}
