% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{PRISMA2geotiff}
\alias{PRISMA2geotiff}
\title{PRISMA2geotiff}
\usage{
PRISMA2geotiff(filepath, overwrite = F, verbose = F)
}
\arguments{
\item{filepath}{character: the filepath to HDF PRISMA dataset}

\item{overwrite}{logical: do you want to overwrite automatically any existing Geotiff file?}

\item{verbose}{logical: lot's of messages if True. Defaults to False.}
}
\value{
logical: TRUE on success or FALSE on error. Writes a geotiff file with the same basename. E.g.
XXX.he5 will be  XXX_VNIR.tif and XXX_SWIR.tif
}
\description{
Exports PRISMA data cube to two geotiff files.
}
\examples{
### filepath<-"/archivio/shared/geodati/raster/OPTICAL/PRISMA/"
### PRS_L2D_STD_20200418101701_20200418101706_0001.he5"
### fn <- PRISMA2geotiff(filepath)
}
