% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{PRISMA2rast}
\alias{PRISMA2rast}
\title{PRISMA2rast}
\usage{
PRISMA2rast(filepath)
}
\arguments{
\item{filepath}{the filepath to HDF PRISMA dataset}
}
\value{
A list object with
\itemize{
\item swir - raster::brick object
\item vnir - raster::brick object
}
}
\description{
PRISMA2rast
}
\examples{
### filepath<-"/archivio/shared/geodati/raster/
## OPTICAL/PRISMA/PRS_L2D_STD_20200418101701_20200418101706_0001.he5"
### fn <- PRISMA2rast(filepath)
}
