% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read.h5}
\alias{read.h5}
\title{Title}
\usage{
read.h5(filepath)
}
\arguments{
\item{filepath}{}
}
\value{

}
\description{
Title
}
\examples{
filepath<-"/archivio/shared/geodati/raster/OPTICAL/PRISMA/PRS_L2D_STD_20200418101701_20200418101706_0001.he5"
fn <- read.h5(filepath)
}
