% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{PRISMA2rast}
\alias{PRISMA2rast}
\title{PRISMA2rast
Reads PRISMA data and saves to a terra::rast object}
\usage{
PRISMA2rast(filepath, verbose = F)
}
\arguments{
\item{filepath}{text of the filepath to HDF PRISMA dataset}

\item{verbose}{boolean - lot's of messages if True}
}
\value{
A list object with
\itemize{
\item swir - terra::rast object
\item vnir - terra::rast object
}
}
\description{
PRISMA2rast
Reads PRISMA data and saves to a terra::rast object
}
