% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{PRISMA2rast}
\alias{PRISMA2rast}
\title{PRISMA2rast}
\usage{
PRISMA2rast(filepath, verbose = F)
}
\arguments{
\item{filepath}{character: text of the filepath to HDF PRISMA dataset}

\item{verbose}{logical: lot's of messages if True. Defaults to False.}
}
\value{
A list of terra::rast objects
\itemize{
\item swir data cube
\item vnir data cube
}
}
\description{
Reads PRISMA data and returns a terra::rast object
}
\examples{
### filepath<-"/archivio/shared/geodati/raster/OPTICAL/PRISMA/"
### filename<-"PRS_L2D_STD_20200418101701_20200418101706_0001.he5"
### fn <- PRISMA2rast( file.path(filepath, filename), verbose=T)
}
